+++ lib/data_forge/dsl/commands.rb
@@ -14,6 +14,12 @@ module DataForge
 
 
 
+      def filter(source, options = {}, &filter_block)
+        Transformation::Filter.from_input(source, options, &filter_block).execute
+      end
+
+
+
       def deduplicate(source, options = {})
         Transformation::Deduplication.from_input(source, options).execute
       end