+++ lib/data_forge/transformation.rb
@@ -2,6 +2,7 @@ module DataForge
   module Transformation
 
     autoload :Deduplication, 'data_forge/transformation/deduplication'
+    autoload :Filter, 'data_forge/transformation/filter'
     autoload :RubyTransformation, 'data_forge/transformation/ruby_transformation'
     autoload :RubyTransformationContext, 'data_forge/transformation/ruby_transformation_context'
     autoload :TransformationBase, 'data_forge/transformation/transformation_base'