+++ CHANGELOG.md
@@ -4,6 +4,7 @@ In reverse chronological order:
 
 ## 0.1.2 (in progress)
 
+* Added the `filter` transformation to filter records of a file source with a boolean-yielding block.
 * Added the `trash` command to delete obsolete (or temporary) file sources.
 
 ## 0.1.1